+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support for all transactions relevant from protocol version 7 and onwards. This includes expanding `AccountTransactionPayload`
   with the necessary variants, and corresponding constructor functions for `AccountTransaction`.
+- `WalletConnectSendTransactionParam` and `WalletConnectSignMessageParam` for decoding parameters received with walletconnect requests.
 
 ## [0.1.1] - 2024-04-29